store UserPreferences properly (#27)
- use androidx datastore api to store preferences (Mode and last
location)
- update some gradle plugins
- fix and improve first compose test a little
- change from float to double for lat/lon values
- fix hardcoded city name in ui
- reload weather data when user preferences changed

+++ app/src/main/java/com/jonas/hillitsweather/data/remote/WeatherDto.kt
@@ -6,10 +6,10 @@ import com.squareup.moshi.Json
 class WeatherDataDto {
 
     @field:Json(name = "lon")
-    var lon: Float = 0.toFloat()
+    var lon: Double = 0.toDouble()
 
     @field:Json(name = "lat")
-    var lat: Float = 0.toFloat()
+    var lat: Double = 0.toDouble()
 
     @field:Json(name = "timezone")
     var timezone: String? = null
